third_party/WebKit/LayoutTests/fast/dom/shadow/tabstop-property-expected.txt

This is basic behavior test for tabStop attribute.

Testing tabStop property and attribute
Test tabStop normal assignment behavior
PASS div.tabStop is false
PASS div.tabStop is false
PASS div.tabStop is false
Test tabStop override by tabindex property
PASS div.tabStop is true
PASS div.tabStop is false
PASS div.tabStop is true
Test tabStop override by tabindex attribute
PASS div.tabStop is true
PASS div.tabStop is false
PASS div.tabStop is true
Test tabStop change after tabindex change
PASS div.tabStop is false
PASS div.tabStop is true
Test tabStop change before tabindex change which will be overridden
PASS div.tabStop is true
PASS div.tabStop is false
Test finished.
PASS successfullyParsed is true

TEST COMPLETE
